Refactored and rebased
