+++ src/components/Home.tsx
@@ -3,9 +3,8 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 
 export default function HomePage() {
   return (
-    <React.Fragment>
-      <CssBaseline />
+    <div>
       Home :)
-    </React.Fragment>
+    </div>
   );
 }
